requirements.txt
hdfs
python-dotenv
